draw a nice dark line between app, lib, and include files
Now we have a single header file that defines the API between the application
code and MDNS daemon.  We also have a single header file that defines the
interface between mdns and the os.  The API isn't very good, but the line is
drawn.

Also rolled some coding style cleanups and moved to just use stdint.h

git-svn-id: https://cozybit.svn.cvsdude.com/zeroconf/trunk@91 6676a129-3127-0410-961e-573e41a41fa5