do not include dns question in dns response unless necessary
I noticed that our mdns responder was not responding to avahi.  Log messages
report that the responses were invalid.  By experimentation, it appears that
avahi does not like that we were including the original question in the
response.  However, the python DNS resolver that we use for testing requires
it.  I verified this last bit by reading the python source.  Finally, as a
responder, avahi seems to selectively include the question or not depending on
the src port of the dns message.  I emulated this behavior so that ping would
work again.

git-svn-id: https://cozybit.svn.cvsdude.com/zeroconf/trunk@97 6676a129-3127-0410-961e-573e41a41fa5