adapt mdns_os.h to be a more complete porting layer
In our original vision, we imagined that the porter would have to write a few
os-specific functions and that would be all.  But to support various compiler
features (like packed data structures) and system features (like stdio.h
features, string.h) with sufficient flexibility, we need a more elaborate
header.