Added mdns sniffer
The mdns sniffer was created to sniff for mdns packets specific
to our mdns tests.  Use the sniffer class is mdns_tool.sniffer, other
parts are intended to be used by the sniffer.

This also adds to the configuration file and to README for documentation.

The tests have had a test added to test the sniffer.