eliminate legacy support for queries from DNS (i.e., not mDNS) clients
The python dns package does not understand the "cache flush" bit of mDNS
message.  Also, it expects response packets to contain the questions to which
they are referring, which is not supported in mDNS.  So we eliminate support
for this in the mdns library, and make our cloned copy of python dns mDNS
friendly.