Add random wait between 0-250ms before first probe
The specification, in section 8.1, requires that implmentations wait a random
time between 0 and 250ms before sending the first probe in order to avoid
a case where a group of devices might send synchronized probes.  This patch
adds the random delay before the statemachine.

In order for this to happen, it introduces a new required function that must
be present in the port.  This is documented in mdns_port.h for the function
mdns_rand_range().  An appropriate implmentation has been added to the
Linux example.