Adjust struct.pack format string for cross-platform use
On my Mac, the struct.pack format string of 'L' refers to an 8 byte
value (causing an error), while on the Linux, it is a 4 byte value.
Replacing this format with 'I', makes it cross-platform compatable
as 'I' is 4 bytes on both platforms.  This makes the tests work
properly on my system while retaining the use on the Linux box.