Add check for response packet size
Response packets might have different content than a probe packet
so we construct and test the size. My stratagy for this was to create
a query and then process that into a response.

I included a test case, however I haven't been able to find a manual
test that passes the probe size, but fails the the response size
check.  I'm fairly sure the test isn't testing what I want it to.