Change sniffer to global start to correct random test failures
The sniffer now is setup to start 1 thread only as a global thread
and the start routine is simplfied to clearing the queue.  I also
changed a few default timeouts. This solves the cases where random
failures abound due to the sniffer state.  However, AnswerAllSRVProbes
always fails due to a sync issue between the test and the uut.  This
is a test-specific issue that needs to be looked at.