Add random delay before responding to queries
This changes the state machine to delay certain types of responses
for between 20 and 120 msec by adding a new state and some new
logic to the (formerly) IDLE state.

The return values of mdns_prepare_response were also changed to
convey information back to the caller so it is known if a delay
is necessary.