refactor querier to use revised algorithm
When messages come in, there are generally going to be very few records, but
there may be many service instances that we're monitoring.  So if we can save
ourselves from searching the list of service instances and search the list of
records instead, we'll be faster and scale more reasonably with increasing
cache size.