fix support for multiple record discovery
We were resetting the arec after putting it in the active list.  This was
leading to some orphaned arecs, which was causing crashes.  We were also
assuming that we would _always_ receive a PTR record for a service before
receiving its SRV record.  This is not always the case, and was causing some
crashes.